Finalize wrapper, all converter test succeeding
